requirements.txt
aiohttp==1.0.5
astroid==1.6.3
async-timeout==3.0.0
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
click==6.7
discord.py==0.16.12
Flask==1.0.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
multidict==4.3.1
oauthlib==2.0.7
pycodestyle==2.4.0
pylint==1.8.4
six==1.11.0
twitter==1.18.0
urllib3==1.22
websockets==3.4
Werkzeug==0.14.1
wrapt==1.10.11
